requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via pytest
certifi==2019.11.28
chardet==3.0.4
click==7.0                # via code-annotations
code-annotations==0.3.3
configparser==4.0.2       # via importlib-metadata
contextlib2==0.6.0.post1  # via importlib-metadata
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0             # via pytest-cov
djangorestframework==3.9.4
drf-yasg==1.17.0
funcsigs==1.0.2           # via pytest
idna==2.8
importlib-metadata==1.3.0  # via pluggy, pytest
inflection==0.3.1
itypes==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
more-itertools==5.0.0     # via pytest, zipp
packaging==19.2
pathlib2==2.3.5           # via importlib-metadata, pytest, pytest-django
pbr==5.4.4                # via stevedore
pluggy==0.13.1            # via pytest
py==1.8.0                 # via pytest
pyparsing==2.4.5
pytest-cov==2.8.1
pytest-django==3.7.0
pytest==4.6.7             # via pytest-cov, pytest-django
python-slugify==4.0.0     # via code-annotations
pytz==2019.3
pyyaml==5.2               # via code-annotations
requests==2.22.0
ruamel.ordereddict==0.4.14 ; python_version == "2.7"
ruamel.yaml.clib==0.2.0
ruamel.yaml==0.16.5
scandir==1.10.0           # via pathlib2
six==1.13.0
stevedore==1.31.0         # via code-annotations
text-unidecode==1.3       # via python-slugify
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata
